requirements.txt
flask
tensorflow
numpy
pillow
scikit-learn
joblib
